requirements.txt
fastmcp
git+https://github.com/SonnyLabs/sonnylabs_py.git
python-dotenv
